+++ praktikumsplaner-frontend/src/plugins/index.ts
@@ -1,8 +1,8 @@
 import type { App } from "vue";
 
 import i18n from "@/plugins/i18n";
-import router from "../router";
 import pinia from "./pinia";
+import router from "./router.ts";
 import vuetify from "./vuetify";
 
 export function registerPlugins(app: App) {
